% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{setioswogrid}
\alias{setioswogrid}
\title{Sets a grid of SS3 runs for the IOTC SWO OM}
\usage{
setioswogrid(scenarios, cpues, dir = paste0("grid_", format(Sys.time(),
  "\%Y\%m\%d")), base = system.file("ext-data/sa", package = "iotcswomse"),
  name = "swo", from = 1, write = TRUE)
}
\description{
Sets a grid of SS3 runs for the IOTC SWO OM
}
\examples{
base <- '../inst/ext-data/sa/'
name <- 'swo'
from <- 1
write <- TRUE

}
