% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{setioswogrid}
\alias{setioswogrid}
\title{Sets a grid of SS3 runs for the IOTC SWO OM}
\usage{
setioswogrid(scenarios, cpues, dir = paste0("grid_", format(Sys.time(),
  "\%Y\%m\%d")), base = system.file("ext-data/sa", package = "ioswomse"),
  name = "swo", from = 1, write = TRUE)
}
\description{
Sets a grid of SS3 runs for the IOTC SWO OM
}
\examples{
# Simple example with a single parameter
scenarios  <- list(steepness=0.7, 0.8)
setioswogrid(scenarios, dir=tempdir())
}
